refactor warningmessage
